# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para consulta de temperatura por CEP",
        "title": "Weather API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@weather-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/weather/{zipcode}": {
            "get": {
                "description": "Retorna a temperatura atual em Celsius, Fahrenheit e Kelvin para um CEP brasileiro",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Obter temperatura por CEP",
                "parameters": [
                    {
                        "type": "string",
                        "example": "01310-100",
                        "description": "CEP brasileiro (8 dígitos)",
                        "name": "zipcode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Temperatura atual",
                        "schema": {
                            "$ref": "#/definitions/domain.Temperature"
                        }
                    },
                    "404": {
                        "description": "CEP não encontrado",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "CEP inválido",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Temperature": {
            "type": "object",
            "properties": {
                "temp_C": {
                    "type": "number"
                },
                "temp_F": {
                    "type": "number"
                },
                "temp_K": {
                    "type": "number"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}